+++ sdk/spring/spring-cloud-azure-service/src/main/java/com/azure/spring/service/storage/queue/QueueServiceClientProperties.java
@@ -16,4 +16,6 @@ public interface QueueServiceClientProperties extends StorageProperties {
 
     QueueMessageEncoding getMessageEncoding();
 
+    Boolean getCreateQueueIfNotExists();
+
 }