+++ sdk/spring/spring-integration-azure-storage-queue/CHANGELOG.md
@@ -5,7 +5,10 @@
 ### Features Added
 - Support StorageQueueMessageConverter as a bean to support customize ObjectMapper.
 ### Breaking Changes
-
+- Remove `StorageQueueOperation`.
+- Deprecate configure checkpoint mode for StorageQueueTemplate, and support only MANUAL mode.
+- Remove auto creating Storage Queue when send/receive messages via `StorageQueueTemplate`, and provide a property of
+  `spring.cloud.azure.storage.queue.create-queue-if-not-exists` to support enabling the auto creation.
 ### Bugs Fixed
 
 ### Other Changes